requirements.txt
asttokens==2.2.1
backcall==0.2.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.8
decorator==5.1.1
et-xmlfile==1.1.0
executing==1.2.0
Flask==2.3.3
Flask-MySQLdb==1.0.1
fonttools==4.42.0
idna==3.4
ipykernel==6.25.0
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mysql-connector-python==8.1.0
mysqlclient==2.2.0
nest-asyncio==1.5.7
numpy==1.25.2
opencv-python==4.8.0.74
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==4.21.12
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.15.1
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
pyzmq==25.1.0
requests==2.31.0
six==1.16.0
stack-data==0.6.2
tornado==6.3.2
traitlets==5.9.0
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.6
Werkzeug==2.3.7
